pull keeper replication out of ReplicatedAccessStorage for shared use
