+++ src/Access/IAccessStorage.h
@@ -217,6 +217,14 @@ public:
     virtual void backup(BackupEntriesCollector & backup_entries_collector, const String & data_path_in_backup, AccessEntityType type) const;
     virtual void restoreFromBackup(RestorerFromBackup & restorer, const String & data_path_in_backup);
 
+    // Static versions for use in non-member functions (ZookeeperReplicator)
+    [[noreturn]] static void throwIDCollisionCannotInsert(
+        const UUID & id, AccessEntityType type, const String & name, AccessEntityType existing_type, const String & existing_name, const String & storage_name);
+    [[noreturn]] static void throwNameCollisionCannotInsert(AccessEntityType type, const String & name, const String & storage_name);
+    [[noreturn]] static void throwNameCollisionCannotRename(AccessEntityType type, const String & old_name, const String & new_name, const String & storage_name);
+    [[noreturn]] static void throwNotFound(const UUID & id, const String & storage_name);
+    [[noreturn]] static void throwNotFound(AccessEntityType type, const String & name, const String & storage_name);
+
 protected:
     virtual std::optional<UUID> findImpl(AccessEntityType type, const String & name) const = 0;
     virtual std::vector<UUID> findAllImpl(AccessEntityType type) const = 0;
@@ -247,13 +255,7 @@ protected:
     static String formatEntityTypeWithName(AccessEntityType type, const String & name) { return AccessEntityTypeInfo::get(type).formatEntityNameWithType(name); }
     static void clearConflictsInEntitiesList(std::vector<std::pair<UUID, AccessEntityPtr>> & entities, LoggerPtr log_);
     virtual bool acquireReplicatedRestore(RestorerFromBackup &) const { return false; }
-    [[noreturn]] void throwNotFound(const UUID & id) const;
-    [[noreturn]] void throwNotFound(AccessEntityType type, const String & name) const;
     [[noreturn]] static void throwBadCast(const UUID & id, AccessEntityType type, const String & name, AccessEntityType required_type);
-    [[noreturn]] void throwIDCollisionCannotInsert(
-    const UUID & id, AccessEntityType type, const String & name, AccessEntityType existing_type, const String & existing_name) const;
-    [[noreturn]] void throwNameCollisionCannotInsert(AccessEntityType type, const String & name) const;
-    [[noreturn]] void throwNameCollisionCannotRename(AccessEntityType type, const String & old_name, const String & new_name) const;
     [[noreturn]] void throwReadonlyCannotInsert(AccessEntityType type, const String & name) const;
     [[noreturn]] void throwReadonlyCannotUpdate(AccessEntityType type, const String & name) const;
     [[noreturn]] void throwReadonlyCannotRemove(AccessEntityType type, const String & name) const;
@@ -303,7 +305,7 @@ std::shared_ptr<const EntityClassT> IAccessStorage::read(const String & name, bo
     if (auto id = find<EntityClassT>(name))
         return read<EntityClassT>(*id, throw_if_not_exists);
     if (throw_if_not_exists)
-        throwNotFound(EntityClassT::TYPE, name);
+        throwNotFound(EntityClassT::TYPE, name, storage_name);
     else
         return nullptr;
 }
